[CWUtils]     1. Fix some compile errors     2. Use enum class instead of enum     3. Use std::min/max instead of min/max
[WppLog]
    1. Update Window WPP generation script to search WPP macro in *.h
files
    2. Re-define log levels
    3. Add FuncEnter() and FuncLeave() macro

[CWString]
    1. Rename GetStringTypeA() to GuessStringTypeA() because collision
with Win32 API

[CWSharedMem]
    1. Add GetName() to get shared memory name
    2. Add CFixedTypeShmQueue to support general shared memory use case

+++ C++/Windows/CWTime.cpp
@@ -229,7 +229,7 @@ UINT64 CStopWatch::Get1QIntervalInNano()
         return 0;
     }
 
-    INT nFirstQuartilePos = m_vecAllInt.size() / 4;
+    size_t nFirstQuartilePos = m_vecAllInt.size() / 4;
     std::nth_element( m_vecAllInt.begin(), m_vecAllInt.begin() + nFirstQuartilePos, m_vecAllInt.end() );
     return m_vecAllInt[nFirstQuartilePos];
 }
@@ -259,7 +259,7 @@ UINT64 CStopWatch::Get3QIntervalInNano()
         return 0;
     }
 
-    INT nThirdQuartilePos = m_vecAllInt.size() * 3 / 4;
+    size_t nThirdQuartilePos = m_vecAllInt.size() * 3 / 4;
     std::nth_element( m_vecAllInt.begin(), m_vecAllInt.begin() + nThirdQuartilePos, m_vecAllInt.end() );
     return m_vecAllInt[nThirdQuartilePos];
 }
